app/vmselect/promql: move tenant filter cache to a separate file
